Initialize project and first steps
My first intuition was to represent the knight's
possible moves with an object which also holds
methods for each of the possible moves and a method
to check whether the move is valid with respect to
the boundaries of the chessboard.

I then took my first steps with the knightMoves()
function and decided to tackle the task with a
Breadth-First-Search algorithm. Thereby I quickly
was able to have the function reach the target square
on the chessboard. It remains to be solved, how to
capture the shortest path as a collection of moves
to be taken. Once solved I easily can derive the
shortest necessary number of steps from the array's
length.