Finish and test project
I finished the project by finding solutions to the
problems I was facing earlier, where I was not able
to trace the path.

I solved this by storing each visited vertex with
the vertex it came from in a 'visited' array. Once
the function reaches the target square and breaks
the loop, I am able to trace the path back, starting
from the target square.